discord-seedbox-bot/requirements.txt
discord.py==2.4.0
paramiko==3.4.0
python-dotenv==1.0.1
